Updated to Java commit 8b84957 (2010.06.17): Change = to include type of boxed numbers. Use == for inter-type numeric equivalence. Plus: fixed some return typing problems objmethod.

+++ Clojure/Clojure/CljCompiler/Ast/NumberExpr.cs
@@ -96,7 +96,7 @@ namespace clojure.lang.CljCompiler.Ast
             Type t = _n.GetType();
 
             if (t == typeof(int))
-                return Expression.Constant((long)_n, typeof(long));
+                return Expression.Constant((long)(int)_n, typeof(long));
             else if (t == typeof(double))
                 return Expression.Constant((double)_n, typeof(double));
             else if ( t == typeof(long) )
